+++ R/builtin-future.apply.R
@@ -1,34 +1,39 @@
-# lapply(X = xs, FUN = FUN, ...) =>
+# future.apply::future_lapply(X = xs, FUN = FUN, ...) =>
 #
 # local(
-#   lapply(X = (.progressr_along <- xs), FUN = function(..., .progressr_progressor) {
+#   future.apply::future_lapply(X = (.progressr_along <- xs), FUN = function(..., .progressr_progressor) {
 #     on.exit(.progressr_progressor())
 #     FUN(...)
 #   }, .progressr_progressor = progressr::progressor(along = .progressr_along))
 # )
 #
-progressify_base <- local({
+progressify_future.apply <- local({
   function(expr, fcn_name, fcn, ..., envir = parent.frame()) {
     names <- names(expr)
     if (is.null(names)) names <- rep("", length.out = length(expr))
     names <- names[-1]
     target_names <- names(formals(fcn))[seq_along(names)]
     unnamed <- setdiff(target_names, names)
+    ddd <- which(unnamed == "...")
+    if (length(ddd) > 0L) {
+      stopifnot(length(ddd) == 1L)
+      unnamed <- unnamed[seq_len(ddd - 1L)]
+    }
     names[names == ""] <- unnamed
     names <- c("", names)
 
     idx_X <- idx_data <- idx_FUN <- idx_n <- idx_expr <- NULL
 
-    if (fcn_name %in% c("by")) {
+    if (fcn_name %in% c("future_by")) {
       idx_data <- which(names == "data")
       idx_FUN <- which(names == "FUN")
-    } else if (fcn_name %in% c("eapply")) {
+    } else if (fcn_name %in% c("future_eapply")) {
       idx_X <- which(names == "env")
       idx_FUN <- which(names == "FUN")
-    } else if (fcn_name %in% c("lapply", "sapply", "vapply")) {
+    } else if (fcn_name %in% c("future_lapply", "future_sapply", "future_vapply")) {
       idx_X <- which(names == "X")
       idx_FUN <- which(names == "FUN")
-    } else if (fcn_name %in% c("replicate")) {
+    } else if (fcn_name %in% c("future_replicate")) {
       idx_n <- which(names == "n")
       idx_expr <- which(names == "expr")
     } else {
@@ -87,29 +92,29 @@ progressify_base <- local({
     }
 
     bquote(local(.(as.call(parts))))
-  } ## progressify_base()
+  } ## progressify_future.apply()
 })
 
 
-append_builtin_transpilers_for_base <- local({
+append_builtin_transpilers_for_future.apply <- local({
   known_fcns <- list(
-    apply = c,
-    by = c,
-    eapply = c,             ## done
-    lapply = c,             ## done
-    .mapply = c,
-    mapply = c,
-    Map = c,
-    replicate = c,          ## done
-    sapply = c,             ## done
-    tapply = c,
-    vapply = c              ## done
+    future_apply = c,
+    future_by = c,
+    future_eapply = c,             ## done
+    future_lapply = c,             ## done
+    future_.mapply = c,
+    future_mapply = c,
+    future_Map = c,
+    future_replicate = c,          ## done
+    future_sapply = c,             ## done
+    future_tapply = c,
+    future_vapply = c              ## done
   )
 
   template <- bquote_compile(function(expr, options) {
-    ns <- baseenv()
+    ns <- getNamespace("future.apply")
     fcn <- get(.(fcn_name), mode = "function", envir = ns)
-    progressify_base(expr, fcn_name = .(fcn_name), fcn = fcn, envir = parent.frame())
+    progressify_future.apply(expr, fcn_name = .(fcn_name), fcn = fcn, envir = parent.frame())
   })
 
   make_transpiler <- function(fcn_name) {
@@ -118,20 +123,19 @@ append_builtin_transpilers_for_base <- local({
   }
 
   function() {
-    ## base::apply(), ...
+    ## future.apply::future_apply(), ...
     transpilers <- list()
     for (fcn_name in names(known_fcns)) {
       transpilers[[fcn_name]] <- list(
-        label = sprintf("base::%s() transpiler", fcn_name),
+        label = sprintf("future.apply::%s() transpiler", fcn_name),
         transpiler = make_transpiler(fcn_name)
       )
     } ## for (fcn_name ...)
-    transpilers <- list(base = transpilers)
+    transpilers <- list(future.apply = transpilers)
     
     append_transpilers("progressify::built-in", transpilers)
     
     ## Return required packages
-    c("base", "progressr")
+    c("future.apply", "progressr")
   }
 })
-