+++ R/register_all_transpilers.R
@@ -3,4 +3,5 @@
 register_all_transpilers <- function() {
   ## Map-reduce packages (base-R)
   append_builtin_transpilers_for_base()
+  append_builtin_transpilers_for_future.apply()
 } ## register_all_transpilers()